Move RemsService to Finna\Service\RemsService.
